requirements.txt
azure-functions==1.6.0
dnspython==2.1.0
pymongo==3.11.3
